requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
torch==2.2.1
scikit-learn==1.4.1
streamlit-folium>=0.11.0
requests>=2.26.0
python-dateutil>=2.8.0
matplotlib>=3.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0
pillow>=8.3.0
netCDF4>=1.5.7
xarray>=0.19.0
cartopy>=0.19.0 